rename dict model and fix fisrt arg name on test setup
