+++ obspy/core/event/dictionary.py
